Committing changes from Assembla code plus initial working kicking behavior
This is just the code from Assembla that wasn't committed before the
transfer to GitHub. Whenever a player gets an active powerup, it renders
that powerup icon over the Active Button. Sets the player's active
powerup to Kick in the constructor for testing purposes.

Also adding rhe initial beahvior for kicking bombs. If a bomb is in
front of the player (this check is very loose right now, it'll get
tightened heavily later on), it adds 0.2 to the position based on
direction in the update method till the bombs explodes. There's no
collision detection, so it can be kicked through blocks and even off the
screen.