Kicking Friction
Adding a Friction Factor constant to the game that slows the bombs down
gradually after they've been kicked. They'll travel almost the entire
width of the screen before exploding. This can be changed, but I found
that having a value over 0.002 will cause the speed to eventually become
negative and cause the bomb to boomerang back towards the player (might
be a cool effect/powerup later on though!)
Also reset the speed of the bomb when kicked so that they can be kicked
more than once.
Added some early value for the Throw powerup as well.